+++ packages/firebase_auth/firebase_auth/ios/firebase_auth/Sources/firebase_auth/include/Private/PigeonParser.h
@@ -4,14 +4,15 @@
  * BSD-style license that can be found in the LICENSE file.
  */
 
-#if TARGET_OS_OSX
-#import <FirebaseAuth/FirebaseAuth.h>
-#else
-@import FirebaseAuth;
-#endif
-
 #import <Foundation/Foundation.h>
 #import "../Public/firebase_auth_messages.g.h"
+
+@class FIRAuthDataResult;
+@class FIRUser;
+@class FIRActionCodeSettings;
+@class FIRAuthTokenResult;
+@class FIRTOTPSecret;
+@class FIRAuthCredential;
 
 @interface PigeonParser : NSObject